#1 fix: 불필한 설정 파일 삭제
