+++ geulkkoli/src/main/java/com/geulkkoli/TestDataInit.java
@@ -1,9 +1,9 @@
 package com.geulkkoli;
 
-import com.geulkkoli.domain.Post;
-import com.geulkkoli.domain.Users;
-import com.geulkkoli.respository.PostRepository;
-import com.geulkkoli.respository.UsersRepository;
+import com.geulkkoli.domain.post.Post;
+import com.geulkkoli.domain.user.Users;
+import com.geulkkoli.domain.post.PostRepository;
+import com.geulkkoli.domain.user.UsersRepository;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.boot.context.event.ApplicationReadyEvent;
