+++ geulkkoli/src/main/java/com/geulkkoli/domain/post/ImplPostRepository.java
@@ -1,6 +1,5 @@
-package com.geulkkoli.respository;
+package com.geulkkoli.domain.post;
 
-import com.geulkkoli.domain.Post;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Repository;
