+++ geulkkoli/src/main/java/com/geulkkoli/domain/post/PostRepository.java
@@ -1,0 +1,8 @@
+package com.geulkkoli.domain.post;
+
+import java.util.Optional;
+
+public interface PostRepository {
+    Post save(Post post);
+    Optional<Post> findById(Long id);
+}
