+++ geulkkoli/src/main/java/com/geulkkoli/domain/user/UsersRepository.java
@@ -1,0 +1,8 @@
+package com.geulkkoli.domain.user;
+
+import java.util.Optional;
+
+public interface UsersRepository {
+    Users save(Users user);
+    Optional<Users> findById(Long id);
+}
