--- geulkkoli/src/main/java/com/geulkkoli/respository/UsersRepository.java
@@ -1,10 +1,0 @@
-package com.geulkkoli.respository;
-
-import com.geulkkoli.domain.Users;
-
-import java.util.Optional;
-
-public interface UsersRepository {
-    Users save(Users user);
-    Optional<Users> findById(Long id);
-}
