+++ geulkkoli/src/main/java/com/geulkkoli/web/post/PostController.java
@@ -1,7 +1,7 @@
-package com.geulkkoli.seungin.controller;
+package com.geulkkoli.web.post;
 
-import com.geulkkoli.domain.Post;
-import com.geulkkoli.seungin.service.PostService;
+import com.geulkkoli.domain.post.Post;
+import com.geulkkoli.domain.post.PostService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Controller;
@@ -10,9 +10,6 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.ModelAttribute;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
-
-import javax.servlet.http.HttpSession;
-import java.util.List;
 
 @Slf4j
 @Controller
