+++ geulkkoli/src/test/java/com/geulkkoli/respository/PostRepositoryTest.java
@@ -1,6 +1,7 @@
 package com.geulkkoli.respository;
 
-import com.geulkkoli.domain.Post;
+import com.geulkkoli.domain.post.Post;
+import com.geulkkoli.domain.post.PostRepository;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
