+++ geulkkoli/src/test/java/com/geulkkoli/respository/UserRepositoryTest.java
@@ -1,7 +1,7 @@
 package com.geulkkoli.respository;
 
-import com.geulkkoli.domain.Users;
-import org.assertj.core.api.Assertions;
+import com.geulkkoli.domain.user.Users;
+import com.geulkkoli.domain.user.UsersRepository;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
@@ -10,7 +10,6 @@ import org.springframework.transaction.annotation.Transactional;
 import java.util.Optional;
 
 import static org.assertj.core.api.Assertions.*;
-import static org.junit.jupiter.api.Assertions.*;
 
 @SpringBootTest
 @Transactional
